add basic debugger(stepper) feature to krun
git-svn-id: http://k-framework.googlecode.com/svn/trunk@7370 82a86efa-2583-11df-8ea0-49ba94de29ab

Former-commit-id: c67b9205c9f348dc5c5393136bde4d2e73a9dd4c [formerly 8c91dee6526327981353aeb19439ab3818d53c32]
Former-commit-id: 3d4ddc3562139e3841ff646209a24168a7960add

+++ javasources/JKRun/src/ro/uaic/info/fmse/jkrun/FileUtil.java
@@ -122,23 +122,26 @@ public class FileUtil {
 	public static void renameFolder(String oldName, String newName) throws IOException {
 		File srcFile = new File(oldName);
 		boolean bSucceeded = false;
-		try {
-			File destFile = new File(newName);
-			//test if the krun directory is empty and if is not empty delete it
-			if (destFile.exists()) {
-				if (!FileUtil.deleteDirectory(destFile)) {
-					throw new IOException(oldName + " was not successfully deleted");
+		
+		if (srcFile.exists()) {
+			try {
+				File destFile = new File(newName);
+				//test if the krun directory is empty and if is not empty delete it
+				if (destFile.exists()) {
+					if (!FileUtil.deleteDirectory(destFile)) {
+						throw new IOException(oldName + " was not successfully deleted");
+					}
 				}
-			}
-			//rename krun temp directory into "krun" 
-			if (!srcFile.renameTo(destFile)) {
-				throw new IOException(oldName + " was not successfully renamed to " + newName);
-			} else {
-				bSucceeded = true;
-			}
-		} finally {
-			if (bSucceeded) {
-				srcFile.delete();
+				//rename krun temp directory into "krun" 
+				if (!srcFile.renameTo(destFile)) {
+					throw new IOException(oldName + " was not successfully renamed to " + newName);
+				} else {
+					bSucceeded = true;
+				}
+			} finally {
+				if (bSucceeded) {
+					srcFile.delete();
+				}
 			}
 		}
 	}
